services/cluster-collection/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.1.0
dnspython==2.3.0
exceptiongroup==1.1.0
Flask==2.2.2
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
llvmlite==0.39.1
MarkupSafe==2.1.1
mock==5.0.1
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.2
pika==1.3.2
pluggy==1.0.0
py==1.11.0
pymongo==4.3.3
pynndescent==0.5.8
pytest==7.2.1
pytest-cov==4.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
umap-learn==0.5.3
urllib3==1.26.14
watchdog==2.2.1
Werkzeug==2.2.2
zipp==3.11.0
